requirements.txt
Flask==2.3.2
requests
google-generativeai
# No extra dependencies required for web search integration (uses PubChem API and Gemini) 